tapo/requirements.txt
dotenv
tapo
python-telegram-bot
requests